+++ week4/weather/src/components/WeekDetailCardInfo.jsx
@@ -1,74 +1,33 @@
 import React from "react";
-import { useState, useEffect } from "react";
-import { useParams } from "react-router-dom";
 import DetailCardInfo from "./DetailCardInfo";
 import styled from "styled-components";
-
+import WeekDataHooks from "../hooks/weekDataHooks";
+import { useParams } from "react-router-dom";
 /**
  * 주간 날씨 컴포넌트 : week/area 형태로 라우팅 시
  */
 const WeekDetailCardInfo = () => {
-  const [cardListData, setCardListData] = useState();
-  const { area } = useParams();
-  const getFiveDetailCardInfo = (area) => {
-    fetch(
-      `https://api.openweathermap.org/data/2.5/forecast?q=${area}&appid=${
-        import.meta.env.VITE_APP_WEATHER
-      }&units=metric`
-    )
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.cod == 200) {
-          const detailData = [];
-          // 날짜 : 어제, 오늘, 내일, 이틀 뒤, 3일 뒤
-          // 인덱스 : 0, 8, 16, 24, 32 (18:00 기준)
-          for (let i = 0; i < 33; i += 8) {
-            let detailDataTmp = {
-              name: data.city.name, // 지역이름
-              weather: [
-                {
-                  description: data.list[i].weather[0].description, // 날씨 설명
-                },
-              ],
-              main: {
-                temp: data.list[i].main.temp, // 현재 온도
-                feels_like: data.list[i].main.feels_like, // 체감기온
-                temp_min: data.list[i].main.temp_min, // 최저
-                temp_max: data.list[i].main.temp_max, // 최고
-              },
-              clouds: {
-                all: data.list[i].clouds.all, // 구름 %
-              },
-              dt_txt: data.list[i].dt_txt.slice(6, 10), // 날짜
-            };
-            detailData.push(detailDataTmp);
-          }
-          setCardListData(detailData);
-        }
-      })
-      .catch((err) => console.log(err));
-  };
+  const area = useParams();
+  const { cardListData } = WeekDataHooks(area);
 
-  useEffect(() => {
-    getFiveDetailCardInfo(area);
-  }, [area]);
-
-  if (cardListData) {
-    return (
-      <St.CardListWrapper>
-        {cardListData &&
-          cardListData.map((item) => (
-            <DetailCardInfo
-              isDayOrWeek={"week"}
-              cardData={item}
-              key={cardListData.indexOf(item)}
-            />
-          ))}
-      </St.CardListWrapper>
-    );
-  } else {
-    return <div>에러입니다</div>;
-  }
+  return (
+    <>
+      {cardListData ? (
+        <St.CardListWrapper>
+          {cardListData &&
+            cardListData.map((item) => (
+              <DetailCardInfo
+                isDayOrWeek={"week"}
+                cardData={item}
+                key={cardListData.indexOf(item)}
+              />
+            ))}
+        </St.CardListWrapper>
+      ) : (
+        <div>로딩 중입니다</div>
+      )}
+    </>
+  );
 };
 
 export default WeekDetailCardInfo;
